requirements.txt
pyobjc-framework-AVFoundation==12.1
pyobjc-framework-Cocoa==12.1
fastmcp==3.2.4
Pillow==12.2.0